fix: improve the code
