+++ src/services/chat/index.ts
@@ -90,21 +90,6 @@ class ChatService {
 
     let pluginIds = [...(enabledPlugins || [])];
 
-    // Expand 'klavis' to all connected Klavis tool identifiers
-    if (pluginIds.includes('klavis')) {
-      const { getToolStoreState } = await import('@/store/tool');
-      const { klavisStoreSelectors } = await import('@/store/tool/selectors');
-
-      // Remove 'klavis' and add all Klavis tool identifiers
-      pluginIds = pluginIds.filter((id) => id !== 'klavis');
-
-      const toolStoreState = getToolStoreState();
-      const klavisTools = klavisStoreSelectors.klavisAsLobeTools(toolStoreState);
-      const klavisIdentifiers = klavisTools.map((tool) => tool.identifier);
-
-      pluginIds = [...pluginIds, ...klavisIdentifiers];
-    }
-
     const toolsEngine = createAgentToolsEngine({
       model: payload.model,
       provider: payload.provider!,
@@ -115,7 +100,7 @@ class ChatService {
       provider: payload.provider!,
       toolIds: pluginIds,
     });
-    
+
     // ============  2. preprocess messages   ============ //
 
     const agentStoreState = getAgentStoreState();
